preprocessing oci names in validate citations, validate citations is now modular
