+++ config.ini
@@ -57,6 +57,7 @@ identifier=
 [COCI]
 prefix=020
 parser=oc.index.parsing.crossref:CrossrefParser
+validator=oc.index.validate.crossref:CrossrefValidator
 source=https://api.crossref.org/works/[[citing]]
 agent=https://w3id.org/oc/index/prov/pa/1
 baseurl=https://w3id.org/oc/index/coci/
