+++ index/python/src/parsing/crossref.py
@@ -20,6 +20,7 @@ from oc.index.parsing.base import CitationParser
 
 class CrossrefParser(CitationParser):
     def __init__(self):
+        super().__init__()
         self._rows = []
         self._doi_manager = DOIManager()
