+++ index/python/src/parsing/crowdsourced.py
@@ -21,6 +21,7 @@ from oc.index.parsing.base import CitationParser
 
 class CrowdsourcedParser(CitationParser):
     def __init__(self):
+        super().__init__()
         self._rows = []
         self._doi_manager = DOIManager()
