+++ scripts/glob_doci.py
