+++ scripts/glob_noci.py
