+++ scripts/trim_crossref.py
