+++ setup.py
@@ -51,15 +51,15 @@ setup(
     exclude_package_data={"": ["./test_data"]},
     entry_points={
         "console_scripts": [
-            "cnc=oc.index.scripts.cnc:main",
-            "ocds=oc.index.scripts.ocds:main",
-            "glob_crossref=oc.index.scripts.crossref_glob:main",
-            "glob_doci=oc.index.scripts.doci_glob:main",
-            "glob_noci=oc.index.scripts.noci_glob:main",
-            "trim_crossref=oc.index.scripts.crossref_trim:main",
-            "validate_crossref=oc.index.scripts.crossref_validate:main",
-            "checkmetadata_crossref=oc.index.scripts.crossref_checkmetadata:main",
-            "oci=oc.index.scripts.oci:main",
+            "oc.index.cnc=oc.index.scripts.cnc:main",
+            "oc.index.datasource=oc.index.scripts.datasource:main",
+            "oc.index.glob.crossref=oc.index.scripts.glob_crossref:main",
+            "oc.index.glob.doci=oc.index.scripts.glob_doci:main",
+            "oc.index.glob.noci=oc.index.scripts.glob_noci:main",
+            "oc.index.trim_crossref=oc.index.scripts.trim_crossref:main",
+            "oc.index.validate=oc.index.scripts.validate_citations:main",
+            "oc.index.metadata.crossref=oc.index.scripts.metadata_crossref:main",
+            "oc.index.oci=oc.index.scripts.oci:main",
         ],
     },
 )
